Update app.js
Set the production environment variable first
process.env.NODE_ENV = "production";

Log it to see if it works
console.log("N0DE_ENV:", process.env.NODE_ENV);

The aforementioned steps were taken because the
log in production showed that the database url
is being set to development; the database is
refusing connection. If this is the case, the
present code in the config file is not setting
the database url to production.